+++ src/Resources/Resources/ActiveDirectory/Cmdlets/NewAzureADUserCommand.cs
@@ -42,7 +42,7 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
         [ValidateNotNullOrEmpty]
         public string ImmutableId { get; set; }
 
-        [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = "The mail alias for the user.")]
+        [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The mail alias for the user.")]
         public string MailNickname { get; set; }
 
         [Parameter(Mandatory = false, ValueFromPipelineByPropertyName = true, HelpMessage = "It must be specified if the user should change the password on the next successful login. Default behavior is to not change the password on the next successful login.")]
@@ -56,6 +56,7 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
             {
                 AccountEnabled = true,
                 DisplayName = DisplayName,
+                MailNickname = MailNickname,
                 PasswordProfile = new PasswordProfile
                 {
                     Password = decodedPassword,
@@ -63,11 +64,6 @@ namespace Microsoft.Azure.Commands.ActiveDirectory
                 },
                 UserPrincipalName = UserPrincipalName
             };
-
-            if (this.IsParameterBound(c => c.MailNickname))
-            {
-                userCreateparameters.MailNickname = MailNickname;
-            }
 
             if (this.IsParameterBound(c => c.ImmutableId))
             {